requirements/base.txt
djangorestframework
social-auth-app-django
fuzzywuzzy
python-Levenshtein
django-filter
django-storages
boto3
celery
pika
django-celery-results
django-celery-beat

